Create NumberGuessing
In this guessing game, the computer will come up with a random number between 1 and 1000. The
player must then continue to guess numbers until the player guesses the correct number. For every
guess, the computer will either say “Too high” or “Too low”, or then ask for another input. At the end
of the game, the number is revealed along with the number of guesses it took to get the correct
number.